refactor: separating fields into new components
